+++ minik/core.py
@@ -14,12 +14,14 @@
     limitations under the License.
 """
 
-import json
-import traceback
+from contextlib import contextmanager
 from collections import namedtuple, defaultdict
-from minik.constants import CONFIG_ERROR_MSG
+
+from minik.exceptions import MinikViewError
+from minik.models import Response
+from minik.builders import APIGatewayRequestBuilder
 from minik.fields import (update_uri_parameters, cache_custom_route_fields)
-from minik.exceptions import MinikError, MinikViewError
+from minik.middleware import (ServerErrorMiddleware, ExceptionMiddleware, ContentTypeMiddleware)
 from minik.status_codes import codes
 
 SimpleRoute = namedtuple('SimpleRoute', ['view', 'methods'])
@@ -43,7 +45,23 @@ class Minik:
     """
 
     def __init__(self, **kwargs):
+        self._debug = kwargs.get('debug', False)
+
+        self._request_builder = kwargs.get('request_builder', APIGatewayRequestBuilder())
+        self._error_middleware = kwargs.get('server_error_middleware', ServerErrorMiddleware())
+        self._exception_middleware = kwargs.get('exception_middleware', ExceptionMiddleware())
+
         self._routes = defaultdict(list)
+        self._middleware = [
+            ContentTypeMiddleware()
+        ]
+
+    @property
+    def in_debug(self):
+        return self._debug
+
+    def add_middleware(self, middleware_instance):
+        self._middleware.append(middleware_instance)
 
     def get(self, path, **kwargs):
         return self.route(path, methods=['GET'], **kwargs)
@@ -86,35 +104,36 @@ class Minik:
         :param context: The aws context included in every lambda function execution
         """
 
-        request = MinikRequest(event, context)
+        request = self._request_builder.build(event, context)
         self.request = request
+        self.response = Response(
+            status_code=codes.ok,
+            headers={'Content-Type': 'application/json'}
+        )
 
-        try:
-
+        with error_handling(self):
             route = self._find_route(request)
-
             update_uri_parameters(route, request)
-            response = self._execute_view(route.view, request)
+            self._execute_view(route.view)
 
-        except MinikViewError as pe:
-            response = JsonResponse({'error_message': str(pe)}, status_code=pe.status_code)
-        except Exception as te:
-            tracer = ''.join(traceback.format_exc())
-            print(tracer)
-            response = JsonResponse({'error_message': str(te), 'trace': tracer}, status_code=500)
+        # After executing the view run all the middlewares in sequence. If a middleware
+        # fails, handle the exception and move on. This code needs to run after the
+        # execution of the views in its own contenxt given that we do want to run this
+        # even if the view itself raised an exception.
+        with error_handling(self):
+            for middleware in self._middleware:
+                middleware(self)
 
-        return response.to_dict()
+        return self.response.to_dict()
 
-    def _execute_view(self, view, request):
+    def _execute_view(self, view):
         """
-        Given a view function, execute the view with the given uri_parameters as
-        argumetns and return a JsonResponse.
+        Given a view function, execute the view and update the body of the current
+        response.
+        :param view: The function to execute.
         """
 
-        if request.uri_params:
-            return JsonResponse(view(**request.uri_params))
-
-        return JsonResponse(view())
+        self.response.body = view(**self.request.uri_params) if self.request.uri_params else view()
 
     def _find_route(self, request):
         """
@@ -148,73 +167,23 @@ class Minik:
         return target_route[0]
 
 
-class MinikRequest:
+@contextmanager
+def error_handling(minik_app):
     """
-    Simple wrapper of the data object received from API Gateway. This object will
-    parse a given API gateway event and it will transform it into a more user
-    friendly object to operate on. The idea is that a view does not need to be
-    concerned with the inner representation of the APIGateway's event as long as
-    it has access to the underlaying data values in the event.
-    """
+    Context manager used to handle both server side errors and unhandled exceptions.
+    Both cases will be delegated to the appropriate middleware of the minik app
+    instance.
 
-    def __init__(self, event, context):
-
-        headers = self._get_with_default(event, 'headers')
-
-        if 'resource' not in event:
-            raise ConfigurationError(CONFIG_ERROR_MSG)
-
-        self.path = event['path']
-        self.resource = event['resource']
-        self.query_params = self._get_with_default(event, 'queryStringParameters')
-        self.headers = {k.lower(): v for k, v in headers.items()}
-        self.uri_params = event['pathParameters']
-        self.method = event['requestContext']['httpMethod']
-        self._body = event['body']
-        # The parsed JSON from the body. This value should
-        # only be set if the Content-Type header is application/json,
-        # which is the default content type.
-        self._json_body = None
-        self.aws_context = context
-
-    def _get_with_default(self, event, param_name, default={}):
-        return event.get(param_name, {}) or default
-
-    @property
-    def json_body(self):
-        """
-        Lazy loading/parsing of the json payload.
-        """
-        if self.headers.get('content-type', '').startswith('application/json'):
-            if self._json_body is None:
-                self._json_body = json.loads(self._body)
-            return self._json_body
-
-
-class JsonResponse:
-    """
-    A very simple wrapper that defines a valid JsonResponse the APIGateway understands.
-    The object encapsulates the headers, status code and body of a response.
+    :param minik_app: The instance of the minik app framework.
     """
 
-    def __init__(self, body, headers=None, status_code=200):
-        self.body = body
-        self.headers = headers or {}
-        self.status_code = status_code
-
-    def to_dict(self, binary_types=None):
-
-        return {
-            'headers': self.headers,
-            'statusCode': self.status_code,
-            'body': json.dumps(self.body)
-        }
+    try:
+        yield
+    except MinikViewError as mve:
+        minik_app._error_middleware(minik_app, mve)
+    except Exception as e:
+        minik_app._exception_middleware(minik_app, e)
 
 
 class BadRequestError(MinikViewError):
     STATUS_CODE = codes.bad_request
-
-
-class ConfigurationError(MinikError):
-    def __init__(self, error_message, *args, **kwargs):
-        super().__init__(self.__class__.__name__ + ': %s' % error_message)