+++ minik/exceptions.py
@@ -11,3 +11,8 @@ class MinikViewError(MinikError):
     def __init__(self, error_message, *args, **kwargs):
         super().__init__(f'{self.__class__.__name__}: {error_message}')
         self.status_code = kwargs.get('status_code', self.STATUS_CODE)
+
+
+class ConfigurationError(MinikError):
+    def __init__(self, error_message, *args, **kwargs):
+        super().__init__(self.__class__.__name__ + ': %s' % error_message)