+++ minik/structures.py
@@ -28,7 +28,6 @@ class LookupDict(dict):
 
     def __getitem__(self, key):
         # We allow fall-through here, so values default to None
-
         return self.__dict__.get(key, None)
 
     def get(self, key, default=None):